addons: implemented setup and defer API functions
The setup function is run for each addon after all addons are loaded
and the metatables set. This allows addon authors to run code at
startup that requires the metamethods.

The defer function allows an addon to admit that it can't parse a
directory after it has said it can.
This may be useful in cases where the can_parse function has to guess
the validity of the file because doing a full test on each directory
load may be too expensive.