addons: modified defer again to not use recursion
There are no user-facing changes to the functionality of defer.
However, now the deferral is done using a while loop within a single
function instead of calling defer and parse_or_defer on the next parser
in the list. This means that an addon can't mess up defer for other
addons by replacing their local version of the defer or parse_or_defer
functions.