keybinds: make setup function work for both addons and custom keybinds
Previously the keybind setup function would only run if custom-keybinds
were enabled. Since the goal is to allow parsers to load custom keybinds
in the future we want it to run in all cases, but only execute the
correct sections of code when needed.

This may be made more elegant in the future