addons: added version number for addon API
This allows me to safely make breaking changes without worrying about
addons randomly breaking. There is now a system for denoting the
significance of changes.

The Major number denotes a large architectural change, the minor patch
is a change to the behaviour (or removal) of an existing part of the API
,and the patch number denotes a bug fix or non-breaking addition to
the API.