parsing: ensure that every parse operation has a unique coroutine
This reduces the efficiency of appending multiple directories to the
playlist and of recursing through multiple directories.
However from my testing these parses seem to be IO bound,
I didn't do hard measurements, but my anecdotal tests didn't notice a
difference, so it's probably not large enough to matter.

What this does do, is make the guarantee that every parse operation
always has a unique coroutine, which means a unique ID. This opens up
some nice future possibilities, and I believe that is worth the
tiny performance reduction when parsing multiple directories in sequence