windir: initial commit of addon
There are now options on both windows and linux for high performance
asynchronous IO through the windir and ls addons.

Asynchronous IO has many advantages such as being able to cancel a parse
operation while it is running, and not locking up the browser while
items are being added to the playlist.