windir: add support for empty directories
the `dir` command returns an error if the directory is empty,
this includes when the directory or file filters make the results empty

Luckliy the error message is different than when an invalid directory is
entered, so detect the exact error message and set the status directly.

Presumably the `dir` command have any behavioural changes for the
foreseeable future.