api: use explicit string library functions on function arguments
This provides better error messages than the str:method syntax since it
will show the string library function in the stacktrace rather than a
generic 'attempt to call/index nil value' error.

This is now done for all functions provided by the addon API. If there
are multiple string library calls then we only need to do this for the
first, afterwards we can be guaranteed that the variable will be a
string.