custom-keybinds: %j now returns 1 for normal selections
You can already control if a keybind has multiselect behaviour using the
`multiselect` flag, there is no benefit to limiting the %j code. With
this change it is no-longer necessary to create two keybinds to
implement the same behaviour for both a single and multiselection that
wants to use %j.