fix memory leak from readonly cache
Since the readonly tables contain a reference to the original table we
can't make the keys weak, at least not in Lua5.1. In Lua 5.2 weakly
keyed tables act as ephemeron tables, so they would probably work
there. I don't have a Lua5.2 build of mpv, so I can't test, but it might
be worth using ephemeral tables when Lua5.2 is detected.