ass styling: replace font size settings with scaling factors
Instead of directly setting the font size you now set a scaling
multiplier.

The reason for this change is to make it easier to replace the icons
with vector graphics in the future. Modifying the vector graphics to
fit with different body sizes is relatively tricky, however the base
scaling factor of the ass overlay can be easily modiifed by changing the
`res_y` value. This means it will be possible to create a single vector
graphic that works at any size.