root: refactor how the root parser is handled
Before the root parser was configured to accept any directory, but
was not part of the usual parser priority list. Instead the scanning
code had special behaviour for loading the root.

This approach had a number of flaws; it was fragile as extra code
was needed in a variety of places; it made the root parser behaviour
different from the other parsers, opening the door for bugs; and it made
it impossible for an addon to replace the root parser.

The new approach changes the root parser to only accept empty
directories, and it is now part of the standard list of parsers
with a priority of `math.huge` (infinity). Instead the directory
scanning code will redirect the browser to the root directory (``),
which is distinct from the *root parser* that it redirected to before.

An additional side effect of this is that addons that use the
`browse_directory` function, or scripts that use the
`get-directory-contents` script-message can now grab the root by sending
an empty directory string. This works when using the `defer` method
as well.