playlist.lua: add type annotations for all functions
The typing is not quite complete yet. The current algorithm stores
extra state values inside list and item objects, which the current class
definitions do not allow.

Rather than continue to store state values in these structures, it may
be better to use a weak table to associate a list or item with the
state data we are currently storing.