cache: disable by default and add option to enable (#125)
The cache tends to cause more problems than it fixes as most people do
not use a slow enough file system for it to really matter.

Now the cache is disabled by default, and people who enable it can
explicitly read the negatives and read about the Ctrl+r button.

The only downside is that addons may have worse handling for the cache
after this.