windir: manually parse utf16
This removes the LuaJit requirements.

This is still rather experimental, so there may be bugs that appear when
parsing non-spec UTF16 strings (which Windows apparently supports in
filepaths!?).
To address this the parser is set to throw errors on any invalid input.
If an error is thrown the parser will abort and file-browser will
fallback to the default parser.