custom-keybinds: example file more generally useful keybinds
Some people seem to treat this file as a set of defaults despite the
installation instructions explicitly saying not to. To avoid people
shooting themselves in the foot all of the less useful keybinds, and the
keybinds that will fail if used on the wrong OS or without the
required commandline tools, have been removed.